requirements.txt
APScheduler==3.11.1
Flask==3.1.1
google-genai==1.49.0
google-generativeai==0.8.5
gunicorn==23.0.0
line-bot-sdk==3.14.0
matplotlib==3.10.5
openpyxl==3.1.5
pandas==2.3.1
requests==2.32.5
tabulate==0.9.0